% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{binary_program_method_options}
\alias{binary_program_method_options}
\title{Options For Use With the Binary Program Method}
\usage{
binary_program_method_options(
  maxit = 500L,
  infimum.maxit = 100L,
  transport.method = transport_options(),
  epsilon = 0.05,
  OTmaxit = 100L,
  model.size = NULL,
  nvars = NULL,
  tol = 1e-07,
  display.progress = FALSE,
  parallel = NULL,
  algorithm = c("exact", "augmented.lagrangian"),
  solver.options = NULL
)
}
\arguments{
\item{maxit}{The maximum iterations for the optimizer. Default is 500.}

\item{infimum.maxit}{Maximum iterations to alternate binary program and Wasserstein distance calculations}

\item{transport.method}{Method for Wasserstein distance calculation. Should be one the outputs of \code{\link[=transport_options]{transport_options()}}}

\item{epsilon}{A value > 0 for the penalty parameter of if using the Sinkhorn method}

\item{OTmaxit}{The number of iterations to run the Wasserstein distance solvers.}

\item{model.size}{What is the maximum number of coefficients to have in the final model. Default is NULL. If NULL, will find models from the minimum size, 0, to the number of columns in \code{X}.}

\item{nvars}{The number of variables to explore. Should be an integer vector of model sizes. Default is NULL which will explore all models from 1 to \code{model.size}.}

\item{tol}{The tolerance for convergence}

\item{display.progress}{Logical. Should intermediate progress be displayed? TRUE or FALSE. Default is FALSE.}

\item{parallel}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{\link[future:plan]{future::plan()}} to run the computations in parallel instead. A cluster from \code{\link[parallel:makeCluster]{parallel::makeCluster()}} or a number of workers is still accepted for now and is used as the plan for the duration of the call.}

\item{algorithm}{How the exact binary program handles the constraint on the number of coefficients. "exact" (default) enforces it directly and "augmented.lagrangian" uses the augmented Lagrangian binary program of Gu, Ahmed, and Dey (2020) \url{doi:10.1137/19M1271695}, which softens the constraint with a penalty large enough to give the same solution and can be faster with some solvers. Ignored by the "lasso" solver.}

\item{solver.options}{Options to be passed on to the solver. See details}
}
\value{
A list with names corresponding to each argument above.
}
\description{
Options For Use With the Binary Program Method
}
\details{
This function will setup the default arguments used by the binary program method. Of note, for the argument \code{solver.options}, If using the "lasso" solver, you should provide arguments such as "penalty", "nlambda", "lambda.min.ratio", "gamma", and "lambda" in a list. A simple way to do this is to feed the output of the \code{\link[=L1_method_options]{L1_method_options()}} function to the argument \code{solver.options.} This will tell the approximate solver, which uses a lasso method that then will project the parameters back to the \eqn{\{0,1\}} space. For the other solvers, pass solver controls as \code{solver.options = list(control = list(...))}. See \code{\link[scip:scip_control]{scip::scip_control()}} for the options of the "scip" solver, such as \code{time_limit}, the ECOS solver package, \code{\link[ECOSolveR:ecos.control]{ECOSolveR::ecos.control()}}, and the options for the mosek solver, \code{\link[Rmosek:mosek]{Rmosek::mosek()}}.
}
\examples{
binary_program_method_options()
# is using the lasso solver for the binary program method to give an approximate solution
binary_program_method_options(solver.options = L1_method_options(nlambda = 50L))
}
\seealso{
\code{\link[=WpProj]{WpProj()}}
}
