% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.dist.compare.R
\name{combine.dist.compare}
\alias{combine.dist.compare}
\title{Combine distance calculationg from the distCompare function}
\usage{
combine.dist.compare(distances)
}
\arguments{
\item{distances}{}
}
\value{
combined distancecompare objects as returned by distCompare function
}
\description{
Combine distance calculationg from the distCompare function
}
