% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.dist.compare.R
\name{combine.distcompare}
\alias{combine.distcompare}
\title{Combine distance calculations from the distCompare function}
\usage{
combine.distcompare(...)
}
\arguments{
\item{...}{\code{distcompare} objects that are the result of \code{\link[=distCompare]{distCompare()}}}
}
\value{
an object of class \code{combine.distcompare}, the combined \code{distcompare} class objects as returned by \code{\link[=distCompare]{distCompare()}} function
}
\description{
Combine distance calculations from the distCompare function
}
\keyword{internal}
