% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{L1_penalty_options}
\alias{L1_penalty_options}
\title{Recognized L1 Penalties}
\usage{
L1_penalty_options()
}
\value{
A character vector with the possible penalties for L1 methods
}
\description{
Recognized L1 Penalties
}
\examples{
L1_penalty_options()
# [1] "lasso"            "ols"              "mcp"              "elastic.net"      "scad"            
# [6] "mcp.net"          "scad.net"         "grp.lasso"        "grp.lasso.net"    "grp.mcp"         
# [11] "grp.scad"         "grp.mcp.net"      "grp.scad.net"     "sparse.grp.lasso"
}
