\name{limbs-package}
\alias{limbs-package}
\alias{limbs}
\docType{package}
\title{
\packageTitle{limbs}
}
\description{
\packageDescription{limbs}
}
\details{

\packageDESCRIPTION{limbs}
\packageIndices{limbs}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{limbs}

Maintainer: \packageMaintainer{limbs}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
