% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.dist.compare.R
\name{plot,combine_distcompare-method}
\alias{plot,combine_distcompare-method}
\title{Plot 'combine.distcompare' Objects}
\usage{
\S4method{plot}{combine_distcompare}(x, ylim = NULL, ylabs = c(NULL, NULL), facet.group = NULL, ...)
}
\arguments{
\item{x}{\code{combine.distcompare} objects resulting from the \code{\link[=combine.distcompare]{combine.distcompare()}} function}

\item{ylim}{y-axis limits}

\item{ylabs}{y-axis labels}

\item{facet.group}{groups to facet by}

\item{...}{additional plotting parameters like alpha}
}
\value{
An object of class \code{plotcombine}that is a list of various diagnostic plots of the \code{WpProj} objects
}
\description{
Plot 'combine.distcompare' Objects
}
\keyword{internal}
