% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.dist.compare.R
\name{plot.combine.dist.compare}
\alias{plot.combine.dist.compare}
\title{plot combine.dist.compare objects}
\usage{
plot.combine.dist.compare(
  x,
  ylim = NULL,
  ylabs = c(NULL, NULL),
  facet.group = NULL,
  ...
)
}
\arguments{
\item{x}{combine.dist compare objects}

\item{ylim}{y-axis limits}

\item{ylabs}{y-axis labels}

\item{facet.group}{groups to facet by}

\item{...}{additional plotting parameters like alpha}
}
\value{

}
\description{
plot combine.dist.compare objects
}
