% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WP_rsquared.R
\name{plot.WPR2}
\alias{plot.WPR2}
\title{Plot \eqn{W_p R^2} objects}
\usage{
plot.WPR2(
  x,
  xlim = NULL,
  ylim = NULL,
  linesize = 0.5,
  pointsize = 1.5,
  facet.group = NULL,
  ...
)
}
\arguments{
\item{x}{\eqn{W_p R^2} object}

\item{xlim}{x-axis limits}

\item{ylim}{y-axis limits}

\item{linesize}{linesize for \link[ggplot2]{geom_line}}

\item{pointsize}{point size for \link[ggplot2]{geom_point}}

\item{facet.group}{Group to do facet_grid by}

\item{...}{currently no effect}
}
\value{

}
\description{
Plot \eqn{W_p R^2} objects
}
