% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.dist.compare.R
\name{plot_ranks}
\alias{plot_ranks}
\title{Plot the rankings on the combine.dist.compare objects}
\usage{
plot_ranks(distances, ylim = NULL, ylabs = c(NULL, NULL), ...)
}
\arguments{
\item{distances}{combine.dist.compare objects}

\item{ylim}{y-axis limits}

\item{ylabs}{y-axis labels}

\item{...}{additional plot arguments like alpha parameter in ggplot2}
}
\value{
object of plotrank
}
\description{
Plot the rankings on the combine.dist.compare objects
}
