% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.dist.compare.R, R/plot_dist.R
\name{print.plotrank}
\alias{print.plotrank}
\alias{print.plotcompare}
\title{Prints the plotrank object}
\usage{
print.plotrank(x)

print.plotcompare(x)
}
\arguments{
\item{x}{}
}
\value{
plots the plotrank objects
}
\description{
Prints the plotrank object
}
