% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceCompare.R
\name{rank.distCompare}
\alias{rank.distCompare}
\title{ranks distcompare objects}
\usage{
rank.distCompare(distances)
}
\arguments{
\item{distances}{Distcompare object}
}
\value{
ranks of distcompare object
}
\description{
ranks distcompare objects
}
