% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceCompare.R
\name{rank_distcompare}
\alias{rank_distcompare}
\title{Ranks \code{distcompare} Objects}
\usage{
rank_distcompare(distances)
}
\arguments{
\item{distances}{A \code{distcompare} object}
}
\value{
The ranks of a \code{distcompare} object as a list containing slots "predictions" and "parameters".
}
\description{
Ranks \code{distcompare} Objects
}
\keyword{internal}
