% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WpProj-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{transport_options}
\alias{wasserstein}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{approxOT}{\code{\link[approxOT:transport_options]{transport_options()}}, \code{\link[approxOT:wasserstein]{wasserstein()}}}
}}

