\name{SparsePosterior-package}
\alias{SparsePosterior-package}
\alias{SparsePosterior}
\docType{package}
\title{
\packageTitle{SparsePosterior}
}
\description{
\packageDescription{SparsePosterior}
}
\details{

\packageDESCRIPTION{SparsePosterior}
\packageIndices{SparsePosterior}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SparsePosterior}

Maintainer: \packageMaintainer{SparsePosterior}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
