% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{transport_options}
\alias{transport_options}
\title{Available Wasserstein Distance Methods}
\usage{
transport_options()
}
\value{
A character vector of available methods
}
\description{
Available Wasserstein Distance Methods
}
\details{
This function features several methods of calculating approximate optimal transport methods in addition to the exact method. The first is the "sinkhorn" method of Cuturi (2013). The second is the "greenkhorn" method of Altschuler et al. (2017). The third is the Hilbert sorting method of Bernton et al (2017). Then there are two novel approximation methods based on the univariate ranks of each covariate to obtain the average rank "rank", and a method based on the univariate distances for each covariate "univariate.approximation.pwr"
}
\examples{
transport_options()
}
