% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasserstein_distance.R
\name{wasserstein}
\alias{wasserstein}
\title{Calculate Wasserstein distances}
\usage{
wasserstein(
  X,
  Y,
  p = 2,
  ground_p = 2,
  observation.orientation = c("colwise", "rowwise"),
  method = c("exact", "sinkhorn", "greenkhorn", "randkhorn", "gandkhorn", "hilbert",
    "rank", "sinkhorn2", "univariate.approximation", "univariate.approximation.pwr",
    "univariate"),
  ...
)
}
\arguments{
\item{X}{Matrix for first group}

\item{Y}{Matrix for second group}

\item{p}{Powr of the wasserstein distance}

\item{ground_p}{Powr of the distance metric. Usually same as `p`}

\item{observation.orientation}{Are observations unique by rows or colummns? Set with "colwise" or "rowwise"}

\item{method}{"exact", "sinkhorn", "greenkhorn","randkhorn", "gandkhorn", "hilbert"}

\item{...}{additional options for sinkhorn based methods. `epsilon` and `niter` determining the hyperparameters for the negative entropy penalty}
}
\value{

}
\description{
Calculate Wasserstein distances
}
