% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPL1.R
\name{WPL1}
\alias{WPL1}
\title{Title}
\usage{
WPL1(
  X,
  Y = NULL,
  theta = NULL,
  power = 2,
  penalty = c("lasso", "ols", "mcp", "elastic.net", "selection.lasso", "scad",
    "mcp.net", "scad.net", "grp.lasso", "grp.lasso.net", "grp.mcp", "grp.scad",
    "grp.mcp.net", "grp.scad.net", "sparse.grp.lasso"),
  model.size = NULL,
  lambda = numeric(0),
  nlambda = 100L,
  lambda.min.ratio = 1e-04,
  gamma = 1,
  maxit = 500L,
  tol = 1e-07,
  ...
)
}
\arguments{
\item{X}{}

\item{Y}{}

\item{theta}{}

\item{power}{}

\item{penalty}{}

\item{model.size}{}

\item{lambda}{}

\item{nlambda}{}

\item{lambda.min.ratio}{}

\item{gamma}{}

\item{maxit}{}

\item{tol}{}

\item{...}{}
}
\value{

}
\description{
Title
}
