% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WpProj-package.R
\docType{package}
\name{WpProj-package}
\alias{WpProj-package}
\title{WpProj: Linear p-Wasserstein Projections}
\description{
Performs Wasserstein projections from the predictive distributions of any model into the space of predictive distributions of linear models. We utilize L1 penalties to also reduce the complexity of the model space. This package employs the methods as described in Dunipace, Eric and Lorenzo Trippa (2020) \doi{10.48550/arXiv.2012.09999}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ericdunipace/WpProj}
  \item Report bugs at \url{https://github.com/ericdunipace/WpProj/issues}
}

}
\author{
\strong{Maintainer}: Eric Dunipace \email{edunipace@mail.harvard.edu} (\href{https://orcid.org/0000-0001-8909-213X}{ORCID})

Authors:
\itemize{
  \item Eric Dunipace \email{edunipace@mail.harvard.edu} (\href{https://orcid.org/0000-0001-8909-213X}{ORCID})
}

Other contributors:
\itemize{
  \item Clemens Schmid \email{clemens@nevrome.de} (\href{https://orcid.org/0000-0003-3448-5715}{ORCID}) (ETA progres bar is adapted from their code) [contributor]
  \item Espen Bernton ('Hilbert Sort' adapted from their code) [contributor]
  \item Mathieu Gerber ('Hilbert Sort' adapted from their code) [contributor]
  \item Pierre Jacob ('Hilbert Sort' adapted from their code) [contributor]
  \item Bin Dai \email{bdai@uwalumni.com} (W2 projections adapted from their 'OEM' code) [contributor]
  \item Jared Huling \email{jaredhuling@gmail.com} (\href{https://orcid.org/0000-0003-0670-4845}{ORCID}) (W2 projections adapted from their 'OEM' code) [contributor]
  \item Yixuan Qiu (W2 projections adapted from their 'OEM' code) [contributor]
  \item Dominic Schuhmacher \email{dominic.schuhmacher@mathematik.uni-goettingen.de} ('Shortsimplex 'optimal transport method adapted from their code) [contributor]
  \item Nicolas Bonneel ('Network Simplex' algorithm adapted from their code) [contributor]
}

}
\keyword{internal}
