% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.dist.compare.R
\name{combine.dist.compare}
\alias{combine.dist.compare}
\title{Combine distance calculationg from the distCompare function}
\usage{
combine.dist.compare(distances)
}
\arguments{
\item{distances}{distcompar object}
}
\value{
class "combine.dist.compare", the combined combine.distcompare objects as returned by \link{distCompare} function
}
\description{
Combine distance calculationg from the distCompare function
}
