% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WP_rsquared.R
\name{combine.WPR2}
\alias{combine.WPR2}
\title{Combine \eqn{W_p R^2} objects}
\usage{
combine.WPR2(...)
}
\arguments{
\item{...}{List of \eqn{W_p R^2} objects}
}
\value{

}
\description{
Combine \eqn{W_p R^2} objects
}
