% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceCompare.R
\name{distCompare}
\alias{distCompare}
\title{Title}
\usage{
distCompare(
  models,
  target = list(posterior = NULL, mean = NULL),
  p = 2,
  ground_p = 2,
  method = "exact",
  quantity = c("posterior", "mean"),
  parallel = NULL,
  transform = function(x) {     return(x) },
  ...
)
}
\arguments{
\item{models}{Models from WpProj methods}

\item{target}{The target to compare the methods to. Should be a list with slots "posterior" to compare the parameters and "mean" to compare predictions}

\item{p}{The power parameter of the Wasserstein distance}

\item{ground_p}{The power of the distance metric. Typically set the same as `p`}

\item{method}{Which approximation to the Wasserstein distance to use. Should be one of "exact", "sinkhorn", "greenkhorn", "gandkhorn", "randkhorn", or "hilbert".}

\item{quantity}{Should the function target the "posterior" or the "mean". Can choose both.}

\item{parallel}{Parallel backend to use for the `foreach` package}

\item{transform}{Transformation function for the predictions.}

\item{...}{other options passed to the wasserstein distance function}
}
\value{
an object of class `distancecompare` with slots `posterior`, `mean`, and `p`.
}
\description{
Title
}
