% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HC.R
\name{HC}
\alias{HC}
\title{Run the Hahn-Carvalho method}
\usage{
HC(
  X,
  Y = NULL,
  theta,
  family = "gaussian",
  penalty = c("elastic.net", "selection.lasso", "lasso", "ols", "mcp", "scad",
    "mcp.net", "scad.net", "grp.lasso", "grp.lasso.net", "grp.mcp", "grp.scad",
    "grp.mcp.net", "grp.scad.net", "sparse.grp.lasso"),
  method = c("selection.variable", "projection"),
  lambda = numeric(0),
  nlambda = 100L,
  lambda.min.ratio = NULL,
  alpha = 1,
  gamma = 1,
  tau = 0.5,
  groups = numeric(0),
  penalty.factor = NULL,
  group.weights = NULL,
  maxit = 500L,
  tol = 1e-07,
  irls.maxit = 100L,
  irls.tol = 0.001,
  intercept = TRUE
)
}
\arguments{
\item{X}{Covariates}

\item{Y}{Predictions}

\item{theta}{Parameters}

\item{family}{Family for method. See \link[oem]{oem}.}

\item{penalty}{Penalty function. See \link[oem]{oem}.}

\item{method}{Should we run a selection variable methodology or projection?}

\item{lambda}{lambda for lasso. See \link[oem]{oem} for this and all options below}

\item{nlambda}{Number of lambda values.}

\item{lambda.min.ratio}{Minimum lambda ratio for self selected lambda}

\item{alpha}{elastic net mixing.}

\item{gamma}{tuning parameters for SCAD and MCP}

\item{tau}{mixing parameter for sparse group lasso}

\item{groups}{A vector of grouping values}

\item{penalty.factor}{Penalty factor for OEM.}

\item{group.weights}{Weights for groupped lasso}

\item{maxit}{Max iteration for OEM}

\item{tol}{Tolerance for OEM}

\item{irls.maxit}{IRLS max iterations for OEM}

\item{irls.tol}{IRLS tolerance for OEM}

\item{intercept}{Should the method include an intercept}
}
\value{
a `WpProj` object with selected covariates and their values
}
\description{
Runs the Hahn-Carvalho method but adapted to return full distributions
}
