% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dist.R
\name{plot,distcompare-method}
\alias{plot,distcompare-method}
\title{Plot \code{distcompare} Objects}
\usage{
\S4method{plot}{distcompare}(
  x = NULL,
  models = NULL,
  ylim = NULL,
  ylabs = c(NULL, NULL),
  xlab = NULL,
  xlim = NULL,
  linesize = 0.5,
  pointsize = 1.5,
  facet.group = NULL,
  ...
)
}
\arguments{
\item{x}{object of class \code{distcompare}}

\item{models}{Can give list of \code{WpProj} outputs and have them turned into \code{distcompare} object for immediate plotting}

\item{ylim}{Limits on y-axis}

\item{ylabs}{Y-axis labels}

\item{xlab}{X-axis labels}

\item{xlim}{Limits of the x-axis}

\item{linesize}{How big to make the lines?}

\item{pointsize}{How big to make the points?}

\item{facet.group}{Should the plots be turned into a \link[ggplot2]{facet_grid}?}

\item{...}{Additional options for the wasserstein distance if just inputing raw \code{WpProj} models}
}
\value{
A \code{ggplot2} object
}
\description{
Plot \code{distcompare} Objects
}
\keyword{internal}
