% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.dist.compare.R
\name{plot_ranks}
\alias{plot_ranks}
\title{Plot the Rankings of the 'combine.distcompare' Objects}
\usage{
plot_ranks(distances, ylim = NULL, ylabs = c(NULL, NULL), ...)
}
\arguments{
\item{distances}{A \code{combine.distcompare} object resulting from the \code{\link[=combine.distcompare]{combine.distcompare()}} function}

\item{ylim}{y-axis limits}

\item{ylabs}{y-axis labels}

\item{...}{additional plot arguments like alpha parameter in ggplot2}
}
\value{
object of class \code{plotrank} which is a list with slots "parameters" and "predictions"
}
\description{
Plot the Rankings of the 'combine.distcompare' Objects
}
\keyword{internal}
