% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgePlot.R
\name{ridgePlot}
\alias{ridgePlot}
\title{Ridge plots for range of coefficients}
\usage{
ridgePlot(
  fit,
  index = 1,
  minCoef = 1,
  maxCoef = 10,
  scale = 1,
  alpha = 0.5,
  full = NULL,
  transform = function(x) {     x },
  xlab = "Predictions",
  bandwidth = NULL
)
}
\arguments{
\item{fit}{A `WpProj` object of list of `WpProj` objects}

\item{index}{The observation number to select. Can be a vector}

\item{minCoef}{The minimum number of coefficients to use}

\item{maxCoef}{The maximum number of coefficients to use}

\item{scale}{How the densities should be scale}

\item{alpha}{Alpha term from ggplot2 object}

\item{full}{"True" prediction to compare to}

\item{transform}{tranform for predictions}

\item{xlab}{x-axis label}

\item{bandwidth}{Bandwidth for kernel}
}
\value{
a `ggridges` plot
}
\description{
Ridge plots for range of coefficients
}
