% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/W2IP.R
\name{W2IP}
\alias{W2IP}
\title{2-Wasserstein distance selection by Integer Programming}
\usage{
W2IP(
  X,
  Y = NULL,
  theta,
  transport.method = transport_options(),
  model.size = NULL,
  infimum.maxit = 100,
  tol = 1e-07,
  solution.method = c("cone", "lp", "cplex", "gurobi", "mosek"),
  display.progress = FALSE,
  parallel = NULL,
  ...
)
}
\arguments{
\item{X}{Covariates}

\item{Y}{Predictions from arbitrary model}

\item{theta}{Parameters of original linear model. Required}

\item{transport.method}{Method for Wasserstein distance calculation. Should be one of "exact", "sinkhorn", "greenkhorn","randkhorn", "gandkhorn","hilbert"}

\item{model.size}{Maximum number of coefficients in interpretable model}

\item{infimum.maxit}{Maximum iterations to alternate binary program and Wasserstein distance calculation}

\item{tol}{Tolerance for convergence of coefficients}

\item{solution.method}{solver to use. Must be one of "cone","lp", "cplex", "gurobi","mosek"}

\item{display.progress}{Should progress be printed?}

\item{parallel}{foreach back end}

\item{...}{Extra args to wasserstein distance methods}
}
\value{

}
\description{
2-Wasserstein distance selection by Integer Programming
}
