% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasserstein_distance.R
\name{wasserstein}
\alias{wasserstein}
\title{Title}
\usage{
wasserstein(
  X,
  Y,
  p = 2,
  ground_p = 2,
  observation.orientation = c("colwise", "rowwise"),
  method = c("exact", "sinkhorn", "greenkhorn", "randkhorn", "gandkhorn", "hilbert",
    "rank", "sinkhorn2", "univariate.approximation", "univariate.approximation.pwr",
    "univariate"),
  ...
)
}
\arguments{
\item{X}{}

\item{Y}{}

\item{p}{}

\item{ground_p}{}

\item{observation.orientation}{}

\item{method}{}

\item{...}{}
}
\value{

}
\description{
Title
}
