% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPL0.R
\name{WPL0}
\alias{WPL0}
\title{Title}
\usage{
WPL0(
  X,
  Y = NULL,
  theta,
  p = 2,
  ground_p = 2,
  method = c("selection.variable", "projection"),
  transport.method = transport_options(),
  epsilon = 0.05,
  maxit = 100,
  parallel = NULL
)
}
\arguments{
\item{X}{}

\item{Y}{}

\item{theta}{}

\item{p}{}

\item{ground_p}{}

\item{method}{}

\item{transport.method}{}

\item{epsilon}{}

\item{maxit}{}

\item{parallel}{}
}
\value{

}
\description{
Title
}
