% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPL0.R
\name{WPL0}
\alias{WPL0}
\title{p-Wasserstein projections with an L0 penalty}
\usage{
WPL0(
  X,
  Y = NULL,
  theta,
  p = 2,
  ground_p = 2,
  method = c("selection.variable", "projection"),
  transport.method = transport_options(),
  epsilon = 0.05,
  maxit = 100,
  parallel = NULL
)
}
\arguments{
\item{X}{matrix of covariates}

\item{Y}{matrix of predictions}

\item{theta}{optional matrix of coefficients from original model, if relevant}

\item{p}{power of the Wasserstein distance}

\item{ground_p}{power of the distance metric}

\item{method}{One of "selection.variable" or "projection". Methods decide whether covariate matrix in `theta` is preserved ("selection.variable") or if new projections are generated ("projection")}

\item{transport.method}{Method for wasserstein distance. One of "exact", "sinkhorn", "greenkhorn","randkhorn", "gandkhorn","hilbert"}

\item{epsilon}{hyperparameter for sinkhorn iterations}

\item{maxit}{max iteration for sinkhorn iterations}

\item{parallel}{foreach backend}
}
\value{
`WpProj` object
}
\description{
p-Wasserstein projections with an L0 penalty
}
