\name{WpProj-package}
\alias{WpProj-package}
\alias{WpProj}
\docType{package}
\title{
\packageTitle{WpProj}
}
\description{
\packageDescription{WpProj}
}
\details{

\packageDESCRIPTION{WpProj}
\packageIndices{WpProj}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{WpProj}

Maintainer: \packageMaintainer{WpProj}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
