\name{WpProj-package}
\alias{WpProj-package}
\alias{WpProj}
\docType{package}
\title{
\packageTitle{WpProj}
}
\description{
\packageDescription{WpProj}
}
\details{
This package performs Wasserstein projections of linear models into arbitrary prediction spaces. Can be used for model selection or interpretation in Bayesian or Frequentist settings.
}
\author{
\packageAuthor{WpProj}

Maintainer: \packageMaintainer{WpProj}
}
\references{
See \url{https://github.com/ericdunipace/WpProj}
}
\keyword{ package }
\examples{
}
