% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WP_rsquared.R
\name{WPR2}
\alias{WPR2}
\alias{WPR2.matrix}
\alias{WPR2.distcompare}
\alias{WPR2.list}
\title{\eqn{W_p R ^2} function to evaluate performance}
\usage{
WPR2(Y = NULL, nu, p = 2, method = "exact", ...)

\method{WPR2}{matrix}(Y, nu, p = 2, method = "exact", base = NULL, ...)

\method{WPR2}{distcompare}(Y = NULL, nu, ...)

\method{WPR2}{list}(Y, nu, p = 2, method = "exact", base = NULL, ...)
}
\arguments{
\item{Y}{Predictions of interest}

\item{nu}{A matrix of competing predictions, a list of matrices from competing predictions or the result from \code{\link{distCompare}}}

\item{p}{Power of the wasserstein distance}

\item{method}{Method for calculating Wasserstein distance}

\item{...}{Arguments passed to Wasserstein distance calculation. See \code{\link{wasserstein}}}
}
\value{
\eqn{W_p R ^2} values
}
\description{
\eqn{W_p R ^2} function to evaluate performance
}
