% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPsimulatedAnnealing.R
\name{WPSA}
\alias{WPSA}
\title{p-Wasserstein distance projections using simulated annealing}
\usage{
WPSA(
  X,
  Y = NULL,
  theta,
  force = NULL,
  p = 2,
  ground_p = 2,
  model.size = 3,
  iter = 1,
  temps = 1000,
  max.time = 3600,
  const = NULL,
  proposal = proposal.fun,
  options = list(method = c("selection.variable", "scale", "projection"),
    transport.method = transport_options(), energy.distribution = "boltzman",
    cooling.schedule = "Geman-Geman", proposal.method = "covariance", epsilon = 0.05,
    OTmaxit = 100),
  display.progress = FALSE,
  parallel = NULL,
  get.theta = TRUE,
  xtx = NULL,
  xty = NULL
)
}
\arguments{
\item{X}{Covariate vector}

\item{Y}{Predictions}

\item{theta}{optional matrix of parameters for generating predictions}

\item{force}{any covariates to force into the model?}

\item{p}{power of the wassersetin distance}

\item{ground_p}{power of the distance metric. Typically same as `p`}

\item{model.size}{Maximum number of coefficients}

\item{iter}{Maximum number of iterations}

\item{temps}{number of temperatures}

\item{max.time}{maximum time in seconds to run}

\item{const}{maximum value for simulated annealing distance}

\item{proposal}{proposal function. There is a default method but can provide your own with parameters xty, cur, idx, force, d, method}

\item{options}{options for simulated annealing}

\item{display.progress}{}

\item{parallel}{foreach backend}

\item{get.theta}{Should the model save the linear coefficients?}

\item{xtx}{precomputed crossproduct \code{crossprod(X,X)}}

\item{xty}{precomputed \code{crossprod(X, Y)}}
}
\value{

}
\description{
p-Wasserstein distance projections using simulated annealing
}
