% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPsimulatedAnnealing.R
\name{WPSA}
\alias{WPSA}
\title{Title}
\usage{
WPSA(
  X,
  Y = NULL,
  theta,
  force = NULL,
  p = 2,
  ground_p = 2,
  model.size = 3,
  iter = 1,
  temps = 1000,
  max.time = 3600,
  const = NULL,
  proposal = proposal.fun,
  options = list(method = c("selection.variable", "scale", "projection"),
    transport.method = transport_options(), energy.distribution = "boltzman",
    cooling.schedule = "Geman-Geman", proposal.method = "covariance", epsilon = 0.05,
    OTmaxit = 100),
  display.progress = FALSE,
  parallel = NULL,
  get.theta = TRUE,
  xtx = NULL,
  xty = NULL
)
}
\arguments{
\item{X}{}

\item{Y}{}

\item{theta}{}

\item{force}{}

\item{p}{}

\item{ground_p}{}

\item{model.size}{}

\item{iter}{}

\item{temps}{}

\item{max.time}{}

\item{const}{}

\item{proposal}{}

\item{options}{}

\item{display.progress}{}

\item{parallel}{}

\item{get.theta}{}

\item{xtx}{}

\item{xty}{}
}
\value{

}
\description{
Title
}
