% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPstepwise.R
\name{WPSW}
\alias{WPSW}
\title{p-Wasserstein distance linear projections using a stepwise method}
\usage{
WPSW(
  X,
  Y,
  theta,
  force = NULL,
  p = 2,
  ground_p = 2,
  direction = c("backward", "forward"),
  method = c("selection.variable", "scale", "projection"),
  transport.method = transport_options(),
  epsilon = 0.05,
  OTmaxit = 100,
  calc.theta = TRUE,
  model.size = NULL,
  parallel = NULL,
  display.progress = FALSE
)
}
\arguments{
\item{X}{matrix of covariates}

\item{Y}{matrix of predictions}

\item{theta}{optional parameter matrix for selection methods.}

\item{force}{Any covariates to force into the model?}

\item{p}{Power of the wasserstein distance}

\item{ground_p}{Distance metric power. Typically the same as `p`}

\item{direction}{forward or backward selection}

\item{method}{"selection.variable" or "projection}

\item{transport.method}{Method for calculating the wasserstein distance. One of "exact", "sinkhorn", "greenkhorn","randkhorn", "gandkhorn","hilbert"}

\item{epsilon}{hyperparameter if using sinkhorn iterations to approximate OT}

\item{OTmaxit}{maximum number of iterations for the opt?imal transport methods}

\item{calc.theta}{should we get the linear coefficients}

\item{model.size}{Maximum model size}

\item{parallel}{foreach backend}

\item{display.progress}{Display intermediate progress}
}
\value{

}
\description{
p-Wasserstein distance linear projections using a stepwise method
}
