% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPstepwise.R
\name{WPSW}
\alias{WPSW}
\title{Title}
\usage{
WPSW(
  X,
  Y,
  theta,
  force = NULL,
  p = 2,
  ground_p = 2,
  direction = c("backward", "forward"),
  method = c("selection.variable", "scale", "projection"),
  transport.method = transport_options(),
  epsilon = 0.05,
  OTmaxit = 100,
  calc.theta = TRUE,
  model.size = NULL,
  parallel = NULL,
  display.progress = FALSE
)
}
\arguments{
\item{X}{}

\item{Y}{}

\item{theta}{}

\item{force}{}

\item{p}{}

\item{ground_p}{}

\item{direction}{}

\item{method}{}

\item{transport.method}{}

\item{epsilon}{}

\item{OTmaxit}{}

\item{calc.theta}{}

\item{model.size}{}

\item{parallel}{}

\item{display.progress}{}
}
\value{

}
\description{
Title
}
