% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPVI.R
\name{WPVI}
\alias{WPVI}
\title{Title}
\usage{
WPVI(
  X,
  Y,
  theta,
  pred.fun = NULL,
  p = 2,
  ground_p = 2,
  transport.method = c("exact", "sinkhorn", "hilbert", "rank",
    "univariate.approximation.pwr", "univariate.approximation"),
  epsilon = 0.05,
  OTmaxit = 100,
  display.progress = FALSE,
  parallel = NULL
)
}
\arguments{
\item{X}{}

\item{Y}{}

\item{theta}{}

\item{pred.fun}{}

\item{p}{}

\item{ground_p}{}

\item{transport.method}{}

\item{epsilon}{}

\item{OTmaxit}{}

\item{display.progress}{}

\item{parallel}{}
}
\value{

}
\description{
Title
}
