% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPVI.R
\name{WPVI}
\alias{WPVI}
\title{p-Wasserstein variable importance}
\usage{
WPVI(
  X,
  Y,
  theta,
  pred.fun = NULL,
  p = 2,
  ground_p = 2,
  transport.method = transport_options(),
  epsilon = 0.05,
  OTmaxit = 100,
  display.progress = FALSE,
  parallel = NULL
)
}
\arguments{
\item{X}{covariates}

\item{Y}{predictions}

\item{theta}{parameters from prediction}

\item{pred.fun}{prediction function. must take variables x, theta}

\item{p}{Power of Wasserstein power}

\item{ground_p}{Power of distance metric}

\item{transport.method}{Transport methods. One of "exact", "sinkhorn", "greenkhorn","randkhorn", "gandkhorn","hilbert"}

\item{epsilon}{Hyperparameter for sinkhorn iterations}

\item{OTmaxit}{maximum number of iterations for the Wasserstein method}

\item{display.progress}{Display intermediate progress}

\item{parallel}{foreach backend}
}
\value{

}
\description{
p-Wasserstein variable importance
}
